fix: component adjustment
