+++ src/pages/holidays/modals/ComponentDetailModal/types.ts
@@ -2,11 +2,3 @@ export interface ModalContent {
   label: string;
   value: string;
 }
-
-export interface RequestComponentDetailProps {
-  title: string;
-  handleClose: () => void;
-  buttonLabel: string;
-  modalContent: ModalContent[];
-  portalId?: string;
-}
